requirements.txt
langchain
pandas
python-dotenv
tiktoken
jq
beautifulsoup4
openai
faiss-cpu